crm-17: showing toast

+++ app/Livewire/Admin/Users/Delete.php
@@ -7,9 +7,12 @@ use App\Notifications\UserDeletedNotification;
 use Illuminate\Contracts\View\View;
 use Livewire\Attributes\{On, Rule};
 use Livewire\Component;
+use Mary\Traits\Toast;
 
 class Delete extends Component
 {
+    use Toast;
+
     public ?User $user = null;
 
     public bool $modal = false;
@@ -35,9 +38,17 @@ class Delete extends Component
     {
         $this->validate();
 
+        if($this->user->is(auth()->user())) {
+
+            $this->addError('confirmation', "You can't delete yourself brow.");
+
+            return;
+        }
+
         $this->user->delete();
 
         $this->user->notify(new UserDeletedNotification());
+        $this->success('User deleted successfully.');
         $this->dispatch('user::deleted');
         $this->reset('modal');
     }
